split up auth.js
