+++ App.js
@@ -1,5 +1,5 @@
 import React, {Component} from 'react';
-import {AppRegistry, Image} from 'react-native';
+import {AppRegistry} from 'react-native';
 import {createAppContainer, createSwitchNavigator} from 'react-navigation';
 import {createStackNavigator} from 'react-navigation-stack';
 import {createBottomTabNavigator} from 'react-navigation-tabs';
